test: consolidate WhatsApp attachment tests and drop low-value cases
Merge the image and document attachment test files into a single
test_whatsapp_attachments.py organised by attachment type, and split
the WhatsappAttachmentHydrationStage tests into a dedicated
test_whatsapp_attachment_hydration_stage.py.

Drop tests that duplicate behaviour already covered by a sibling test
sharing the same parser code path (Twilio MMS treats image and document
identically; WhatsAppMessage.parse routes image and document through the
same branch). Also drop the "empty caption -> empty body" cases that
only re-verify dict-get semantics.
